requirements.txt
Flask==3.1.1
PyJWT==2.8.0